backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "clerk-backend-api>=3.0.3",
    "fastapi>=0.115.13",
    "openai>=1.88.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
